Fix LLVM warning in previous commit

+++ modules/juce_core/native/juce_mac_Network.mm
@@ -248,7 +248,7 @@ public:
 
     void didSendBodyData (NSInteger totalBytesWritten, NSInteger /*totalBytesExpected*/)
     {
-        latestTotalBytes = totalBytesWritten;
+        latestTotalBytes = static_cast<int> (totalBytesWritten);
     }
 
     void finishedLoading()
